Move logger and quotenv to app_files.
