+++ Files for Humans/manifest.yml
