+++ app_files/meta/timestampedLogger.js
